+++ src/lib/route-data/route-data.interface.ts
@@ -1,4 +1,4 @@
-import { LabelFunction } from './route-data-function-types';
+import { LabelFunction } from './route-data-types';
 
 export interface RouteDataInterface {
     label: string | LabelFunction;
